ml-service/requirements.txt
fastapi
pydantic
joblib
pandas
numpy
scikit-learn
catboost
optuna
matplotlib
seaborn
uvicorn
requests

psycopg2-binary
